Fixed the basename so that files no longer are forced to ".pdf.pdf" and instead will just be ".pdf".
This doesn't solve the *larger* problem of when SCOTUS changes the filename and doesn't use the convention it adhered to through 2013 and 2014. In the event that changes, it may be wise to check for hash values in the repository.

As it is, I decided *not* to do that for 2 reasons:

1. It would slow down the app
2. Due to asynchronous nature of Node, it seemed like it could have unintended consequences.

IN any event, it should fix the false positives problem that's emerged in the last few days.

(Incidentally, the ETAGS are *no longer different* on the SCOTUS site. It's rather inexplicable to me why these false positives happened. It is almost as if the SCOTUS server garbled the ETAGS and then self-corrected. But it is only for a few opinions. I'm making a note to check on these opinions for potential changes in Bound Volumes.)